Add files via upload
testing having multiple branches, on being master and one being testing for code to be reviewed by other group members to ensure it is good before being committed to master. Hopefully I did this right.

Updated proj1.c with some error checking to see if "<" or ">" was the last character inputted. Made it so echo command didn't print any of the four special characters and that echo command could print the word echo. Also tried some stuff involving part 3 on the project instructions, I believe I got "PWD >" working how it is supposed to be.