Add files via upload
commands can now accept multiple parameters to be passed to the getenv funk. 
I also added some code that detects where and how many commands the user typed in. Not really being used right now but I suspect it will be useful in the future. Obviously if it ends up not being needed then we can remove it. That code is lines 143-157