A lot of stuff
put in file identification because echo ~ should print ~ even though that isn't how it works in regular bash. TA said so in recitation
essentially files are tokens that have a "/" in it or are simply ".", "..", "~" 
except for the case where it is a parameter for the echo command
code can now tell the difference between these cases and identifies files with an array called filespots

When forking for execv commands the output of the command now prints before the program prompts the user for the next command. 

Reorganized code and made it atually readable. 
i like format of
if ()
{
}
instead of 
if () {
}
I've been coding a lot so I want it this way, if you want to change it back on your personal version on your laptop go for it but
I think it's hard to read nested statements that way
commented "SECTION: " signifies code that does something completely different

Redid conversion of ~ to $HOME in a filepath. Only converts this if it is a file path 
(won't convert if parameter for echo)

Added part that uses stat() function to check if the file paths are valid or not. 
This method should be used to check if commands are valid. 
Part 5 is essentially using this to check if executable is in /bin/ or in /user/bin/ or any of the other $PATH paths that
are separated by ":" *** STILL need to do that part of part 5 but it is kind of set up for it now ***

convert $HOME and environmental variables like that even when they're in a file path (ex. "ls $HOME/OS") 
This took a few hours wrestling with memory allocation and seg faults. Lots of fun but it works now. 
***PLEASE test a bunch of combinations using the env variables to make sure I actually caught everything ***

using an invalid file path now prints an error message and prompts the user for another command as it should instead of completely 
exiting the code like I had it yesterday

Can identify when a .. is being used in a file path. /.. is invalid and something like ~/hey..low isn't considered the .. operator. 
didn't actually implement the function of the .. yet though. Where the test double dot print is is where the functionality needs to go.